handbook restructuring
